[client] dmabuf: lock framebuffer until DMA finishes
This commit uses a second LGMP queue to hold onto the framebuffer until
the DMA completes, at which point it calls lgmpClientMessageDone to free
the framebuffer.

This may timeout. Graceful resync logic is added for the second LGMP queue
to handle the case when a frame is displayed for longer than LGMP timeout.